restructure visualization dir
