--- src/components/Visualization/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Visualization';
