+++ src/lib/visualization/index.ts
@@ -1,0 +1,2 @@
+export { default as Visualization } from './Visualization';
+export * from './types';
